applied search bar component to resources page.
