+++ client/src/components/searchbar/SearchBar.jsx
@@ -1,0 +1,75 @@
+import { useEffect, useState } from "react";
+
+import {
+    InputGroup,
+    InputLeftElement,
+    Input,
+    Flex,
+    Badge,
+} from "@chakra-ui/react";
+
+import { FaSearch } from "react-icons/fa";
+
+export const SearchBar = ( { onSearch, tags = {}, tagFilter = {}, onTag }) => {
+    const [searchInput, setSearchInput] = useState("");
+
+      const handleEnterKeyDown = async (e) => {
+        if (e.key == "Enter") {
+          onSearch(searchInput);
+        }
+      }
+
+  return (
+    <>
+        <InputGroup>
+            <InputLeftElement pointerEvents="none">
+                <FaSearch color="gray.300" />
+            </InputLeftElement>
+            <Input
+                placeholder="Search"
+                variant="filled"
+                borderRadius="full"
+                borderColor={"gray.300"}
+                bg="white.100"
+                _hover={{ bg: "gray.200" }}
+                _focus={{ bg: "white", borderColor: "gray.300" }}
+                value={searchInput}
+                onChange={(e) => setSearchInput(e.target.value)}
+                onKeyDown={handleEnterKeyDown}
+            />
+        </InputGroup>
+
+      { tags && onTag && (
+        <Flex gap={3}
+              maxWidth="100%" 
+              overflowX="auto"
+              css={{
+                '&::-webkit-scrollbar': {
+                  display: 'none'
+                },
+                '-ms-overflow-style': 'none',
+                'scrollbar-width': 'none'
+              }}
+            >
+              {Object.entries(tags).map(([id, name]) => (
+                <Badge
+                  key={id}
+                  onClick={() => {
+                    console.log("clickity clack")
+                    onTag(id)();
+                  }}
+                  rounded="xl"
+                  px={4}
+                  py={1}
+                  colorScheme={tagFilter[id] ? "green" : "red"}
+                  textTransform="none"
+                  cursor="pointer"
+            >
+                {name}
+            </Badge>
+            ))}
+        </Flex>
+      )}
+    </>
+  )
+}
